create single category component and style it
