+++ src/components/categories/index.js
@@ -1,5 +1,7 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
+
+import Category from '../category';
 import AddNew from '../addButton';
 import Styles from './index.module.css';
 import Form from '../form';
@@ -32,14 +34,15 @@ function Categories() {
     <div className="categories">
       <h1 className={Styles.categoryHeading}>Categories</h1>
       {categories.map((category) => (
-        <h1 key={category.id}>
-          <Link to={`/category/${category.id}`}>{category.title}</Link>
-        </h1>
+        // <h1 key={category.id}>
+        //   <Link to={`/category/${category.id}`}>{category.title}</Link>
+        // </h1>
+        <Category key={category.id} id={category.id} title={category.title} />
       ))}
-      
+
       <div className={Styles.addButton}>
         {showForm ? <Form /> : null}
-        <AddNew text="new category" showForm={displayForm}/>
+        <AddNew text="new category" showForm={displayForm} />
       </div>
     </div>
   );
