+++ src/components/category/index.module.css
@@ -1,20 +1,10 @@
-.categoryName {
-  font-size: 2rem;
-  margin-top: 5px;
-  margin-bottom: 20px;
+.singleCategory {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 
-.addButton {
-  position: relative;
-  bottom: -210px;
-  left: 0px;
-}
-
-.addItemForm {
-  height: 100px;
-  width: 250px;
-  background-color: blue;
-  position: fixed;
-  bottom: 60px;
-  left: 45px;
+.Link {
+  text-decoration: none;
+  color: inherit;
 }
